requirements.txt
asgiref==3.8.1
Authlib==1.3.0
automium==0.2.6
automium_web==0.1.1
base==0.0.0
blinker==1.7.0
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
django-allauth==0.62.1
Flask==3.0.3
Flask-OAuth==0.12
Flask-OAuthlib==0.9.6
google-api-core==2.18.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
oauth2==1.9.0.post1
oauthlib==2.1.0
passlib==1.7.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
